cleanup value converters
refactor std converters

cleanup converters pt2

cleanup pt3

cleanup ptxy

cleanup webreq
