+++ shell/common/api/locker_gin.h
@@ -4,8 +4,8 @@
 
 // TODO(deermichel): remove _gin suffix after removing mate
 
-#ifndef SHELL_COMMON_API_LOCKER_H_
-#define SHELL_COMMON_API_LOCKER_H_
+#ifndef SHELL_COMMON_API_LOCKER_GIN_H_
+#define SHELL_COMMON_API_LOCKER_GIN_H_
 
 #include <memory>
 
@@ -35,4 +35,4 @@ class Locker {
 
 }  // namespace gin
 
-#endif  // SHELL_COMMON_API_LOCKER_H_
+#endif  // SHELL_COMMON_API_LOCKER_GIN_H_
